+++ SBS_Backend/sbs/src/main/java/com/securebanking/sbs/repository/TransactionAuthorizationRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TransactionAuthorizationRepo extends JpaRepository<TransactionAuthorization, Long> {
+public interface TransactionAuthorizationRepo extends JpaRepository<TransactionAuthorization, Integer> {
 
 }
